backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
psutil==5.9.8
mss==9.0.1
Pillow==11.1.0
pygetwindow==0.0.9; sys_platform == "win32"
discord.py==2.3.2
python-dotenv==1.0.1
aiohttp==3.9.5
pydantic==2.7.4
pydantic-settings==2.3.3
numpy==2.2.3
